refactor(blog): 루트에 남아 있던 제네레이터 스크립트를 scripts/로 모은다
빌드 스크립트 대부분은 이미 scripts/에 있는데 generate-rss·generate-sitemap·
sync-posts 세 개만 앱 루트에 남아 있었다. 같은 성격의 파일이 두 곳에 흩어져
있을 이유가 없어 scripts/ 아래로 합친다.

옮기면서 깨지는 경로를 함께 고쳤다.

- 스크립트 안의 상대 import(`./lib/*`, `./domain/*`)를 한 단계씩 올린다
- generate-rss·generate-sitemap이 쓰는 publicDir이 `import.meta.url` 기준이라
  `..`를 끼워 앱 루트의 public/을 계속 가리키게 한다
- sync-posts.mjs의 POSTS_SOURCE_DIR·POSTS_TARGET_DIR도 같은 이유로 한 단계씩
  조정한다
- build-content.ts는 cwd가 앱 루트라 args에 `scripts/` 접두만 붙이면 된다
- contract.test.ts의 `@/generate-sitemap`·`@/generate-rss` alias를
  `@/scripts/...`로 옮긴다
- package.json 테스트 글롭에서 `generate-*.test.ts`와 `sync-posts.test.ts`를
  뺀다. 기존 `scripts/**/*.test.ts`가 그대로 커버한다

sync-posts.test.ts는 스크립트를 import하지 않고 인라인 wrapper를 spawn하는
구조라 손댈 게 없었다.

검증: 세 스크립트를 직접 실행해 산출물이 앱 루트 public/에 그대로 떨어지는
것과(sync-posts가 기존 41개를 unchanged로 인식) 테스트 772건, 타입체크,
린트가 모두 통과하는 것을 확인했다.
