+++ apps/blog/web/scripts/sync-posts.test.ts
